services/datapipeline-consumer/requirements.txt
kafka-python==2.0.2
pytest==8.3.3
